test(pagination): add BDD tests verifying pagination documentation invariants
- Introduce a new common fixtures module for shared BDD test state and types.
- Add comprehensive BDD test suite for pagination documentation invariants in a dedicated test file.
- Refactor existing pagination BDD tests to use the common fixtures module.
- Update feature files and step definitions to align with new BDD test structure.
- Improve readability and maintainability of the pagination BDD tests by removing redundant code and organizing test helper functions.
- Minor docstring and spelling corrections in pagination crate code to improve documentation quality.
